feat(csa-process): detect claude-code death during auto-compaction (#1335)
When claude-code triggers auto-compaction it exits but spawns a grandchild
subprocess that inherits stdout, preventing EOF. The daemon read loop blocked
indefinitely instead of detecting the dead process.

Fix: poll the child PID in every watchdog tick (200ms) via
`/proc/{pid}/stat` on Linux. A grace period of one poll interval (200ms)
filters normal exits where stdout closes naturally within microseconds.
Only fires when the child remains in zombie state with stdout still open for
a full interval — the compaction scenario where a grandchild holds the pipe.

- `tool_liveness::pid_has_exited_or_zombie`: reads /proc/{pid}/stat;
  falls back to kill(0) on non-Linux Unix; conservative false on other targets
- `wait_and_capture_with_idle_timeout`: zombie detection in watchdog tick
  of both (with-stderr) and (without-stderr) branches; overrides exit code
  to non-zero and appends diagnostic to stderr_output and summary
- Tests: compaction-death integration test (zombie within 5s), unit test for
  pid_has_exited_or_zombie live vs dead, and regression for normal exit

Also extract ExecutionResult to lib_execution_result.rs and tool_liveness
tests to tool_liveness_tests.rs to keep both files under the 800-line limit.

Bumps workspace version to 0.1.682.
